move process_zonal_statistics to zonal_stats.py
